Simplify design system owners
This replaces the following repeated list of users with
`@gitlab-org/foundations/design-system`:

- `@annabeldunstone`
- `@clavimoniere`
- `@danmh`
- `@jeldergl`
- `@jmiocene`
- `@markrian`
- `@sdejonge`
- `@seggenberger`
- `@thutterer`

At the time of writing, the members of
`@gitlab-org/foundations/design-system` are:

- `@annabeldunstone`
- `@clavimoniere`
- `@chrismicek`
- `@danmh`
- `@jeldergl`
- `@jmiocene`
- `@markrian`
- `@paintedbicycle-gitlab`
- `@seggenberger`
- `@sdejonge`
- `@thutterer`

Which leaves `@tauriedavis @rayana @aregnery @pedroms` as the only
remaining named users.

The net effect of this replacement is to *add* `@paintedbicycle-gitlab`
and `@chrismicek` to the code owners for these files, which I think is
reasonable.