Re-expand design system group
It turns out the group is not currently invited to the project, so it is
not eligible. See
https://docs.gitlab.com/user/project/codeowners/reference/#group-eligibility.

This means it's ignored, which was discovered in
https://gitlab.com/gitlab-org/gitlab-services/design.gitlab.com/-/merge_requests/5197#note_2827410787.

This re-expands the group to its direct members, as a quick workaround.

We should revert this once we have the group invited to the project. We
can't do *that* until we have an owner of the project available, and
there are surprisingly few of those.