Fix /svgs Pages site in production
In production, the base of the `/svgs` Nuxt site was incorrectly set to
`/` because `CI_ENVIRONMENT_URL` isn't set there. It *is* set for review
apps, which is why review apps still work.

So, the fix is to always rely on `CI_PAGES_URL` instead, which is always
correct.

This also removes the unnecessary reference to `CURRENT_PAGES_URL`.