Fix /svgs search parameter
Before, searching for a string removed the `/svgs` path when adding the
`?q=...` search parameter.

This was due to the `$route.path`, since it does not include the
`baseDir` configuration value: it's the path *from* that. So,
`$route.path` was `/` (or `/file_icons`, etc.) rather than `/svgs` (or
`/svgs/file_icons`, etc.).

Instead, this now relies only on the current `location`.