Revert "Bump node version"
This reverts commit 5a9cf680b4d37c1f6b25fbe37ece460c98e378c0, and
ensures a newer `npm` version is available for [trusted publishing][2].

As described in
https://gitlab.com/gitlab-org/gitlab-services/design.gitlab.com/-/merge_requests/5262,
CI jobs which use the Playwright images were broken due to
https://gitlab.com/gitlab-org/gitlab-services/design.gitlab.com/-/merge_requests/5258.

The reason is that the `phantomjs-prebuilt` package (needed by
`svg-sprite` from `packages/gitlab-svgs`) needs `bzip2` to be
installed, but apparently that's not installed in the Playwright images.

The reason those jobs used to succeed was that the `node_modules` cache
prepared by previous jobs using the `node:22.20.0-trixie` images (which
do have `bzip2` installed) can be reused as-is [without rebuilding][1].
This is because both images use Node 22.x.

By updating other images to `node:24.10.0`, that cache isn't useable by
the Playwright image, forcing a rebuild in jobs using it, causing the
`bzip2` failure from above.

So, this restores the utility of the `node_modules` cache between the
images.

The original motivation for the bump to Node 24.x was to get a newer
`npm` version, so trusted publishing would work. Instead, we directly
install a newer `npm` version in the publishing job.

[1]: https://nodejs.org/en/learn/modules/abi-stability#abi-stability-in-nodejs
[2]: https://docs.npmjs.com/trusted-publishers