Include package jobs with pipeline:include-* labels
Sometimes it's useful to run the jobs defined in workspace packages
without changing any of their files.

Now, the `pipeline:include-svgs`, `pipeline:include-ui`, and
`pipeline:include-tw` labels are available to do just that.

Add the labels to a merge request, then trigger a new pipeline, and the
respective jobs will be included in the pipeline.