Bump cypress/browsers image
We were seeing some flakiness in
https://gitlab.com/gitlab-org/gitlab-services/design.gitlab.com/-/merge_requests/5403#note_2938946786,
where Cypress was failing to start, with this error:

```
Failed to connect to socket /run/dbus/system_bus_socket: No such file or
directory
```

Maybe this is fixed in newer images? There is
https://github.com/cypress-io/cypress/issues/32290, but that only
affects versions of Cypress around v15, and we were still on v14.5.4.

Now that we're on v15.7.1, let's try bumping the image anyway and see
what happens.