Fix Storybook publishing
The previous approach fetched the [artifacts by reference name][1],
which:

> Download[s] a job’s artifacts archive from the latest successful
> pipeline using the reference name.

This means that it would always fail, since the parent pipeline wouldn't
yet have completed by the time the child pipeline is triggered. The fix
will be to download from the specific job (by id) directly.

See
https://gitlab.com/gitlab-org/frontend/design-system-pages/-/merge_requests/1
for the corresponding change in the child project.

[1]: https://docs.gitlab.com/api/job_artifacts/#download-job-artifacts-by-reference-name